requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==24.0.0
idna==3.11
oauthlib==3.3.1
packaging==26.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
